Release v3.0.0 with a stabilised public API
- Do not expose internal driver implementation
