+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rv8803"
-version = "2.0.0"
+version = "3.0.0"
 authors = ["Michael de Silva <michael@cyberdynea.io>"]
 edition = "2021"
 repository = "https://github.com/bsodmike/rv8803-rs"
