+++ src/rtc.rs
@@ -1,4 +1,4 @@
-use crate::Rv8803;
+use crate::driver::Driver;
 use chrono::{DateTime, Utc};
 use core::marker::PhantomData;
 #[allow(unused_imports)]
@@ -7,7 +7,7 @@ use defmt::error;
 use heapless::String;
 use shared_bus::BusManager;
 
-/// The [`RTClock`] type is the interface for communicating with the `rv8803` rtc clock chip via a shared bus over I2C.
+/// The [`RTClock`] type is the interface for communicating with the `Driver` rtc clock chip via a shared bus over I2C.
 #[allow(dead_code)]
 pub struct RTClock<'a, I2C, I2cErr, M> {
     datetime: Option<DateTime<Utc>>,
@@ -25,7 +25,7 @@ where
     SharedBusMutex: shared_bus::BusMutex,
     <SharedBusMutex as shared_bus::BusMutex>::Bus: embedded_hal_0_2::blocking::i2c::Write<Error = I2cErr>
         + embedded_hal_0_2::blocking::i2c::WriteRead<Error = I2cErr>,
-    I2cErr: defmt::Format,
+    I2cErr: defmt::Format + core::convert::From<core::num::TryFromIntError>,
 {
     /// Creates a new [`RTClock`].
     pub fn new(bus: &'a BusManager<SharedBusMutex>, address: &u8) -> Self {
@@ -38,16 +38,42 @@ where
         }
     }
 
-    /// Access the driver for the rtc clock.
-    #[must_use]
-    pub fn rtc(&self) -> Rv8803<crate::prelude::Bus<'_, shared_bus::I2cProxy<'_, SharedBusMutex>>> {
+    /// Set time on the Driver module
+    ///
+    /// # Errors
+    ///
+    /// Read/write errors during communication with the `Driver` chip will return an error.
+    #[allow(clippy::too_many_arguments)]
+    pub fn set_time(
+        &mut self,
+        sec: u8,
+        min: u8,
+        hour: u8,
+        weekday: u8,
+        date: u8,
+        month: u8,
+        year: u16,
+    ) -> Result<bool, I2cErr> {
         let proxy = self.bus.acquire_i2c();
+        let mut driver = Driver::from_i2c(proxy, self.device_address);
 
-        Rv8803::from_i2c(proxy, self.device_address)
+        driver.set_time(sec, min, hour, weekday, date, month, year)
+    }
+
+    /// Fetch time from the RTC clock and store it in the buffer `dest`.
+    ///
+    /// # Errors
+    ///
+    /// Read/write errors during communication with the `Driver` chip will return an error.
+    pub fn update_time(&mut self, dest: &mut [u8]) -> Result<bool, I2cErr> {
+        let proxy = self.bus.acquire_i2c();
+        let mut driver = Driver::from_i2c(proxy, self.device_address);
+
+        driver.update_time(dest)
     }
 }
 
-/// The [`RTClock`] type is the interface for communicating with the `rv8803` rtc clock chip directly over I2C.
+/// The [`RTClock`] type is the interface for communicating with the `Driver` rtc clock chip directly over I2C.
 #[allow(dead_code)]
 pub struct RTClockDirect<'a, I2C, I2cErr> {
     datetime: Option<DateTime<Utc>>,
@@ -74,7 +100,7 @@ where
     }
 
     /// Access the driver for the rtc clock.
-    pub fn rtc(self) -> Rv8803<crate::prelude::Bus<'static, I2C>> {
-        Rv8803::from_i2c(self.periph, self.device_address)
+    pub fn rtc(self) -> Driver<crate::prelude::Bus<'static, I2C>> {
+        Driver::from_i2c(self.periph, self.device_address)
     }
 }
